Move public methods to LeanplumInbox
